+++ src/tui/flatten.rs
@@ -142,7 +142,7 @@ impl App {
                 kind: FlatRowKind::HeapNode {
                     node_ordinal: Some(ordinal),
                     distance: Some(snap.node_distance(ordinal)),
-                    shallow_size: snap.node_self_size(ordinal) as f64,
+                    shallow_size: snap.node_self_size(ordinal) as u64,
                     retained_size: snap.node_retained_size(ordinal),
                     reachable_size: self.reachable_sizes.get(&ordinal).copied(),
                     detachedness: Some(detachedness),
